# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Gym App API
  version: 0.1.0
  description: API inicial para el primer vertical de entrenamiento.
servers:
  - url: /api/v1
paths:
  /health:
    get:
      operationId: getHealth
      summary: Comprueba que el servicio está disponible.
      security: []
      responses:
        '200':
          description: Servicio disponible.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    const: ok
  /auth/register:
    post:
      operationId: register
      summary: Crea una cuenta de atleta.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Cuenta creada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          description: El correo ya está registrado.
  /auth/login:
    post:
      operationId: login
      summary: Inicia sesión con correo y contraseña.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Sesión iniciada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Credenciales inválidas.
  /me:
    get:
      operationId: getCurrentUser
      summary: Obtiene la cuenta autenticada.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Cuenta autenticada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: Bearer token ausente o inválido.
  /me/training-profile:
    put:
      operationId: saveTrainingProfile
      summary: Guarda el perfil de entrenamiento.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrainingProfile'
      responses:
        '200':
          description: Perfil guardado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingProfile'
        '401':
          description: Token requerido.
        '422':
          description: Perfil inválido.
    get:
      operationId: getTrainingProfile
      summary: Consulta el perfil de entrenamiento.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Perfil guardado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingProfile'
        '401':
          description: Token requerido.
        '404':
          description: Perfil no creado.
  /workout-plans:
    post:
      operationId: createWorkoutPlan
      summary: Crea un plan manual de entrenamiento.
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateWorkoutPlanRequest'}
      responses:
        '201': {description: Plan creado, content: {application/json: {schema: {$ref: '#/components/schemas/IdResponse'}}}}
        '401': {description: Token requerido.}
        '422': {description: Plan o ejercicio inválido para el perfil.}
  /workout-plans/{planId}/sessions:
    post:
      operationId: createWorkoutSession
      summary: Registra una sesión completada del plan.
      security: [{bearerAuth: []}]
      parameters:
        - name: planId
          in: path
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateWorkoutSessionRequest'}
      responses:
        '201': {description: Sesión registrada, content: {application/json: {schema: {$ref: '#/components/schemas/IdResponse'}}}}
        '401': {description: Token requerido.}
        '403': {description: El plan pertenece a otro usuario.}
        '422': {description: Series inválidas.}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, acceptedTermsAt]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        acceptedTermsAt:
          type: string
          format: date-time
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    AuthResponse:
      type: object
      required: [userId, accessToken]
      properties:
        userId:
          type: string
          format: uuid
        accessToken:
          type: string
    MeResponse:
      type: object
      required: [userId, email]
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
          format: email
    TrainingProfile:
      type: object
      required: [experienceLevel, primaryProfile, secondaryProfiles, goal, availabilityBand, availableDaysPerWeek, sessionDurationMinutes]
      properties:
        experienceLevel:
          type: string
          enum: [BEGINNER, INTERMEDIATE, ADVANCED]
        primaryProfile:
          $ref: '#/components/schemas/TrainingProfileCode'
        secondaryProfiles:
          type: array
          maxItems: 2
          uniqueItems: true
          items:
            $ref: '#/components/schemas/TrainingProfileCode'
        goal:
          type: string
          minLength: 1
        availabilityBand:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        availableDaysPerWeek:
          type: integer
          minimum: 1
          maximum: 7
        sessionDurationMinutes:
          type: integer
          minimum: 15
          maximum: 240
    TrainingProfileCode:
      type: string
      enum: [GENERAL_FITNESS, BODYBUILDING, POWERLIFTING, RUNNING, CROSSFIT, CALISTHENICS]
    IdResponse:
      type: object
      required: [id]
      properties: {id: {type: string, format: uuid}}
    ExercisePlanRequest:
      type: object
      required: [exerciseId, sets, minRepetitions, maxRepetitions]
      properties:
        exerciseId: {type: string, format: uuid}
        sets: {type: integer, minimum: 1}
        minRepetitions: {type: integer, minimum: 1}
        maxRepetitions: {type: integer, minimum: 1}
    WorkoutDayRequest:
      type: object
      required: [name, exercises]
      properties:
        name: {type: string, minLength: 1}
        exercises: {type: array, minItems: 1, items: {$ref: '#/components/schemas/ExercisePlanRequest'}}
    CreateWorkoutPlanRequest:
      type: object
      required: [name, days]
      properties:
        name: {type: string, minLength: 1}
        days: {type: array, minItems: 1, items: {$ref: '#/components/schemas/WorkoutDayRequest'}}
    SetLogRequest:
      type: object
      required: [exerciseId, repetitions]
      properties:
        exerciseId: {type: string, format: uuid}
        repetitions: {type: integer, minimum: 1}
    CreateWorkoutSessionRequest:
      type: object
      required: [sets]
      properties:
        sets: {type: array, minItems: 1, items: {$ref: '#/components/schemas/SetLogRequest'}}